Fix issue #53 / Correct fix for issue #49
The previous fix for issue #49 (Scanning extra files messes up verification) was supposed to ignore an extra file that is identical with the target file during the recovery process. But that fix caused the first extra file with a target file name (without path) matching a substring in the full path of the extra file name to be ignored during recovery attempts (issue #53). This fix discards the extra file only if the  'canonical' target path is identical with the 'canonical' extra file path. Notice that return value of 'ExtraFile::FileName()' is already a canonical full path (see initialization of extra file list). Because all file names in the list of extra files are unique the search for a matching file name can be aborted after the first match.