Fix for #80 ("-n<n>" when "-r<c><n>").
When "-r<c><n>" had been implemented, a check had been added disallowing
to specify filecount on command line when redundancy percentage is not
enforced ("-r0") with the comment "Allow -r0" (04af1d). As this check
does not match the comment, it might be superfluous, but as it breaks
specifying filecount on command line, it has been removed.

A helper function (Par2Creator::ComputeRecoveryBlockCount) internally
calls first ComputeRecoveryFileCount() to calculate recoveryfilecount
(unless specified on command line), but then reset it to zero in case
"-r<c><n>" has been used, breaking "-n<n>". This has been removed as well
as it seemed to make no sense. This was implemented along with the orignal
implementation "specify absolute size of recovery file data" (046044)
and no specific change, so probably just not fully tested.