DANGEROUS: Moved ifdefs from par2cmdline.h into par2lib.h
This was done to make par2lib.h independent of par2cmdline.h.
The code in commandline.cpp still depends on diskfile.h and
stricmp from par2cmdline.h, but that's okay.  The important
part is that other code that calls functions in par2lib.h
do not need to include par2cmdline.h.

Hmmm... there is a dependancy on automake's config.h, but
I think that's okay.