DANGEROUS: Changed WIN32 to _WIN32 or _MSC_VER
WIN32 is not the standard define for Windows code, _WIN32 is.
I made the change while testing cross-compiling with MinGW
(a.k.a., Mingw32).  That is a port of gcc's compiler to target
windows binaries.  I got it to work on Ubuntu by installing
"mingw-w64" and then running:
    ./configure --host=x86_64-w64-mingw32
But, for MinGW, the types do not change.  So it does not define
__int64, for example.  #ifdefs that protected those types were
changed to _MSC_VER, which is defined for Microsoft's Visual C++
compiler.

SQUASH THIS - added .gitignore to ignore .exe files