DANGEROUS, BUG FIX:  Fixed most things found by cppcheck
cppcheck is a static code checker.  It finds uninitialized
variables, unused code, etc..  I fixed most of the stuff
it found.  I didn't fix:
   C-style pointer casts
   using i++ instead of ++i
There were a few issues that it found that I did not fix
because I thought the code was clearer.

We might want to make these changes too, just so it is
easy to run cppcheck and verify everything is okay.