Add par2 test with existing empty files
Created a par2 set with parpar since par2cmdline normally should not
allow empty files to be included in the parset.

This test is currently failing because the test-a.data file is empty.

```
Repair is required.
1 file(s) exist but are damaged.
10 file(s) are ok.
You have 1029 out of 1029 data blocks available.
You have 52 recovery blocks available.
Repair is possible.
You have an excess of 52 recovery blocks.
None of the recovery blocks will be used for the repair.

Wrote 0 bytes to diska

Verifying repaired files:

Target: "test-a.data" - empty.
Repair Failed.
```

Tickets: #128
Signed-off-by: BlackEagle <ike.devolder@gmail.com>