When a sourcefile is detected don't bail on it being 0 bytes
This is a workaround for empty files in a parset breaking repair.

The empty file will always land in the unverifiable sourcefile list.
But we can match it lateron. Currently I don't think we can get it in
the verifiedsourcefiles because there are a ton of guardrails against
loading 0 byte files.

It also looks on first glance this workaround is needed to be able to
verify parfiles created with other tools not skipping on 0 byte files.

Tickets: #128
Signed-off-by: BlackEagle <ike.devolder@gmail.com>